+++ HISTORY.md
@@ -7,6 +7,7 @@
 - Added support for prereleases (`bump --new-prerelease/--finalize-prerelease`) with the ability to specify in the configuration the JS code for bumping the prerelease part
 - The config file keeps track also of the latest stable (non-prerelease) version
 - Added support for `search` and `replace` in config
+- Added support for `stable_only` in config
 
 ### Changed